Add text to nav and title sections and image
Add filler H1 text which will be replaced by the logo.
Add display text for the nav items.

Add an H1, a paragraph, and a button element to the title section.
Insert display text for each new element.

Insert a filler image for the image that will be in the title section.

Add notes to the header, title, options, and testimonial sections to
inform what will go in those sections in the future.